test(e2e): harness-restore — absorb 8 lifecycle/persistence probes
New shared harness for probes that genuinely need multi-launch / pre-launch
fixtures / process-level mutations. Each absorbed case manages its own
electron lifecycle (skipLaunch:true) because the harness-runner #335
capability set does not yet expose the allocated user-data dir to the case
body — and "two launches into the SAME dir" is the whole point of the
restore probes.

Absorbed (8 probes -> 1 harness file, 8 cases):
  - restore                       (2-launch + JSONL fixture)
  - restore-journey-plan          (2-launch + JSONL fixture)
  - restore-journey-question      (2-launch + JSONL fixture)
  - sidebar-resize                (2-launch + persistence)
  - sidebar-rename-dnd-state      (3-launch + persistence)
  - db-corruption-recovery        (pre-launch FS write to ccsm.db)
  - notify-fallback               (pre-launch node_modules rename + sandboxed HOME)
  - import-session                (pre-launch HOME sandboxing + multi-fixture plant)

Reclassifications (NOT moved by this PR — separate worker will absorb into
their correct harness):
  - restore-session-undo, restore-group-undo  -> harness-agent (single-launch)
  - askuserquestion-full                      -> harness-agent (single-launch)
  - sidebar-journey-create-delete             -> harness-agent (single-launch)
  - streaming-partial-frames                  -> harness-real-cli (needs claude bin)

Probes named in planning but absent from tree (no action):
  - restore-journey-permission, tray (no source file ever existed)

Verification:
  $ node scripts/harness-restore.mjs    # 8/8 PASS, 76.7s wall
  $ node scripts/run-all-e2e.mjs        # not regressed (8 probes -> 1 harness)
